env param and move around files
