+++ src/app/tartle/completed/page.tsx
@@ -1,0 +1,18 @@
+/* eslint-disable @next/next/no-sync-scripts */
+import Head from 'next/head';
+import Script from 'next/script';
+
+export default function Completed() {
+  return (
+    <>
+      <Script src='https://source.tartle.co/scripts/endlink-completion.js' />
+      <div className='mt-28 w-screen'>
+        <div className='w-min mx-auto'>
+          <div className='text-4xl tracking-tight font-extrabold   sm:text-5xl md:text-6xl lg:text-7xl'>
+            <div className='text-gray-800 dark:text-gray-200 mb-24'>👍</div>
+          </div>
+        </div>
+      </div>
+    </>
+  );
+}
